requirements.txt
mysql-connector-python==8.0.33
tabulate==0.9.0
inquirer==3.1.3
openpyxl==3.1.2
python-telegram-bot==20.7
schedule==1.2.0
